+++ PSFront/Public/Message Templates/Update-FrontMessageTemplate.ps1
@@ -1,0 +1,47 @@
+function Update-FrontMessageTemplate {
+    <#
+    .SYNOPSIS
+        Short description
+    .DESCRIPTION
+        Long description
+    .EXAMPLE
+        PS C:\> <example usage>
+        Explanation of what the example does
+    .INPUTS
+        Inputs (if any)
+    .OUTPUTS
+        Output (if any)
+    #>
+    [CmdletBinding()]
+    param (
+        [Parameter(ParameterSetName="ById")]
+        [String]$Id,
+
+        [Parameter(ParameterSetName="ByFolderId")]
+        [String]$FolderId,
+
+        [Parameter(Mandatory)]
+        [SecureString]$ApiKey
+    )
+
+    $Params = @{
+        Method   = "GET"
+        ApiKey   = $ApiKey
+    }
+
+    switch ($PSCmdlet.ParameterSetName) {
+        "ById" {
+            $Params["Endpoint"] = "message_templates"
+            $Params["Path"]     = $id
+        }
+        "ByFolderId" {
+            $Params["Endpoint"] = "message_template_folders"
+            $Params["Path"]     = "{0}/message_templates" -f $FolderId
+        }
+        default {
+            $Params["Endpoint"] = "message_templates"
+        }
+    }
+
+    InvokeFrontRestMethod @Params
+}