+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.6] - 2021-11-22
 ### Added
-- New functions: `Update-FrontConversation`, `Find-FrontConversation`, `New-FrontComment`
+- New functions: `Update-FrontConversation`, `Find-FrontConversation`, `New-FrontComment`, `Get-FrontMessageTemplate`, `Get-FrontMessageTemplateFolder`, `Update-FrontMessageTemplate`
 
 ### Changed
 - Renamed parameter `-Tag` to `-TagId` in `Add-FrontConversationTag`